requirements.txt
pandas>=2.0.0
jinja2>=3.1.0
python-dotenv>=1.0.0
pytest>=7.4.0
